backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose
psycopg2-binary
email-validator
python-multipart
python-dotenv
schedule
